requirements.txt
asgiref==3.7.1
Django==4.2.1
django-crispy-forms==2.0
django-storages==1.11.1
Pillow==9.5.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
